+++ create_unet.py
@@ -1,5 +1,4 @@
-import unet
 import models
 
-model = unet.Unet_resnext50()
+model = models.Unet_resnext50()
 models.saveModel(model, 'URX_resnext50', weights_only=True)
